dags/medallion_pipeline_dag.py
"""
medallion_pipeline_dag.py
Airflow DAG — NYC Taxi Medallion Lakehouse Pipeline

Orchestrates Bronze → Silver → Gold Databricks notebooks with:
- Dependency-enforced task ordering
- Retry / backoff policy per layer
- SLA alerts via email callback
- XCom-based row-count tracking for downstream observability
- Automated Great Expectations checkpoint run after Silver
"""

from __future__ import annotations

import json
import logging
from datetime import datetime, timedelta

from airflow import DAG
from airflow.models import Variable
from airflow.operators.empty import EmptyOperator
from airflow.operators.python import BranchPythonOperator, PythonOperator
from airflow.providers.databricks.hooks.databricks import DatabricksHook
from airflow.providers.databricks.operators.databricks import DatabricksRunNowOperator
from airflow.utils.dates import days_ago
from airflow.utils.trigger_rule import TriggerRule

logger = logging.getLogger(__name__)

# ─── Constants & Airflow Variables ───────────────────────────────────────────

DATABRICKS_CONN_ID = "databricks_default"
CATALOG = Variable.get("nyc_taxi_catalog", default_var="main")
DATABASE = Variable.get("nyc_taxi_database", default_var="nyc_taxi")
BATCH_DATE = "{{ ds }}"  # Airflow logical date (YYYY-MM-DD)

# Notebook job IDs pre-registered in Databricks — set via Airflow Variables
BRONZE_JOB_ID = int(Variable.get("bronze_job_id", default_var="1001"))
SILVER_JOB_ID = int(Variable.get("silver_job_id", default_var="1002"))
GOLD_JOB_ID = int(Variable.get("gold_job_id", default_var="1003"))
GE_JOB_ID = int(Variable.get("ge_job_id", default_var="1004"))

ALERT_EMAILS = Variable.get(
    "alert_emails", default_var="charan.neelam@company.com"
).split(",")

# ─── Default Arguments ───────────────────────────────────────────────────────

default_args = {
    "owner": "charan.neelam",
    "depends_on_past": False,
    "email": ALERT_EMAILS,
    "email_on_failure": True,
    "email_on_retry": False,
    "retries": 2,
    "retry_delay": timedelta(minutes=5),
    "retry_exponential_backoff": True,
    "max_retry_delay": timedelta(minutes=30),
}

# ─── SLA Miss Callback ───────────────────────────────────────────────────────


def sla_miss_callback(dag, task_list, blocking_task_list, slas, blocking_tis):
    logger.error(
        "SLA MISSED | DAG: %s | Tasks: %s | Date: %s",
        dag.dag_id,
        task_list,
        datetime.utcnow().isoformat(),
    )


# ─── Python Callables ────────────────────────────────────────────────────────


def check_source_data(**context) -> str:
    """
    Pre-flight check: verify source CSV files exist in cloud storage.
    Returns branch task_id to proceed or skip ingestion.
    """
    DatabricksHook(databricks_conn_id=DATABRICKS_CONN_ID)
    # In production, list DBFS / ADLS path via hook.  Here we use a lightweight
    # SQL check via Databricks SQL endpoint (simplified for illustration).
    logger.info("Pre-flight source data check passed.")
    return "bronze_ingestion"  # always proceed; replace with real path check


def validate_bronze_output(**context) -> None:
    """Pull Bronze row count from XCom and assert minimum threshold."""
    ti = context["ti"]
    run_output = ti.xcom_pull(task_ids="bronze_ingestion", key="notebook_output")
    if run_output:
        try:
            result = json.loads(run_output.replace("'", '"'))
            row_count = result.get("row_count", 0)
            logger.info("Bronze row count: %s", row_count)
            if row_count < 1000:
                raise ValueError(
                    f"Bronze row count {row_count} below minimum threshold of 1,000."
                )
        except (json.JSONDecodeError, KeyError) as exc:
            logger.warning("Could not parse Bronze output: %s", exc)


def validate_silver_output(**context) -> None:
    """Assert Silver rejection rate is within acceptable bounds."""
    ti = context["ti"]
    run_output = ti.xcom_pull(task_ids="silver_transform", key="notebook_output")
    if run_output:
        try:
            result = json.loads(run_output.replace("'", '"'))
            rejection_rate = result.get("rejection_rate_pct", 0)
            logger.info("Silver rejection rate: %.2f%%", rejection_rate)
            if rejection_rate > 20.0:
                raise ValueError(
                    f"Silver rejection rate {rejection_rate:.2f}% exceeds 20% threshold."
                )
        except (json.JSONDecodeError, KeyError) as exc:
            logger.warning("Could not parse Silver output: %s", exc)


def log_pipeline_success(**context) -> None:
    logger.info(
        "Medallion pipeline completed successfully for logical date: %s",
        context["ds"],
    )


# ─── Notebook Job Parameters ─────────────────────────────────────────────────


def _bronze_params() -> dict:
    return {
        "catalog": CATALOG,
        "database": DATABASE,
        "bronze_table": "bronze_trips",
        "load_mode": "incremental",
    }


def _silver_params() -> dict:
    return {
        "catalog": CATALOG,
        "database": DATABASE,
        "bronze_table": "bronze_trips",
        "silver_table": "silver_trips",
        "batch_date": BATCH_DATE,
    }


def _gold_params() -> dict:
    return {
        "catalog": CATALOG,
        "database": DATABASE,
        "silver_table": "silver_trips",
        "batch_date": BATCH_DATE,
    }


def _ge_params() -> dict:
    return {
        "suite_name": "silver_trips_suite",
        "batch_date": BATCH_DATE,
    }


# ─── DAG Definition ──────────────────────────────────────────────────────────

with DAG(
    dag_id="medallion_lakehouse_pipeline",
    description="NYC Taxi Bronze → Silver → Gold + GE quality gate",
    default_args=default_args,
    schedule_interval="0 3 * * *",  # 03:00 UTC daily
    start_date=days_ago(1),
    catchup=False,
    max_active_runs=1,
    tags=["medallion", "delta-lake", "nyc-taxi", "pyspark"],
    sla_miss_callback=sla_miss_callback,
    doc_md=__doc__,
) as dag:

    # ── Pre-flight ────────────────────────────────────────────────────────────
    source_check = BranchPythonOperator(
        task_id="source_data_check",
        python_callable=check_source_data,
        sla=timedelta(minutes=10),
    )

    skip_pipeline = EmptyOperator(task_id="skip_pipeline")

    # ── Bronze ────────────────────────────────────────────────────────────────
    bronze_ingestion = DatabricksRunNowOperator(
        task_id="bronze_ingestion",
        databricks_conn_id=DATABRICKS_CONN_ID,
        job_id=BRONZE_JOB_ID,
        notebook_params=_bronze_params(),
        sla=timedelta(hours=1),
        do_xcom_push=True,
    )

    bronze_validation = PythonOperator(
        task_id="bronze_validation",
        python_callable=validate_bronze_output,
    )

    # ── Silver ────────────────────────────────────────────────────────────────
    silver_transform = DatabricksRunNowOperator(
        task_id="silver_transform",
        databricks_conn_id=DATABRICKS_CONN_ID,
        job_id=SILVER_JOB_ID,
        notebook_params=_silver_params(),
        sla=timedelta(hours=2),
        do_xcom_push=True,
    )

    silver_validation = PythonOperator(
        task_id="silver_validation",
        python_callable=validate_silver_output,
    )

    # ── Great Expectations quality gate ──────────────────────────────────────
    great_expectations_check = DatabricksRunNowOperator(
        task_id="great_expectations_check",
        databricks_conn_id=DATABRICKS_CONN_ID,
        job_id=GE_JOB_ID,
        notebook_params=_ge_params(),
        sla=timedelta(minutes=30),
    )

    # ── Gold ──────────────────────────────────────────────────────────────────
    gold_aggregate = DatabricksRunNowOperator(
        task_id="gold_aggregate",
        databricks_conn_id=DATABRICKS_CONN_ID,
        job_id=GOLD_JOB_ID,
        notebook_params=_gold_params(),
        sla=timedelta(hours=1),
        do_xcom_push=True,
    )

    # ── Post-pipeline ─────────────────────────────────────────────────────────
    pipeline_success = PythonOperator(
        task_id="pipeline_success",
        python_callable=log_pipeline_success,
        trigger_rule=TriggerRule.ALL_SUCCESS,
    )

    end = EmptyOperator(
        task_id="end",
        trigger_rule=TriggerRule.NONE_FAILED_MIN_ONE_SUCCESS,
    )

    # ─── Task Dependencies ────────────────────────────────────────────────────
    (source_check >> [bronze_ingestion, skip_pipeline])

    (
        bronze_ingestion
        >> bronze_validation
        >> silver_transform
        >> silver_validation
        >> great_expectations_check
        >> gold_aggregate
        >> pipeline_success
        >> end
    )

    skip_pipeline >> end
