requirements.txt
# ─── Core Data Processing ─────────────────────────────────────────────────────
pyspark==3.5.0
delta-spark==3.1.0

# ─── Orchestration ────────────────────────────────────────────────────────────
apache-airflow==2.8.1
apache-airflow-providers-databricks==6.2.0

# ─── Data Quality ─────────────────────────────────────────────────────────────
great-expectations==0.18.12

# ─── dbt ──────────────────────────────────────────────────────────────────────
dbt-core==1.7.4
dbt-databricks==1.7.4
dbt-snowflake==1.7.4

# ─── Cloud SDKs ───────────────────────────────────────────────────────────────
azure-storage-blob==12.19.0
azure-identity==1.15.0
boto3==1.34.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# ─── Linting / Formatting ─────────────────────────────────────────────────────
ruff==0.1.14
black==24.1.1
isort==5.13.2
bandit[toml]==1.7.7

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.6.0
tenacity==8.2.3
